backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.3
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.11.9
reportlab==4.0.7
